created directories participants and tradings in teatures, slices moved from shared folder store to features/name/slices
