+++ src/app/components/table/table-body/table-body.tsx
@@ -3,7 +3,7 @@ import { TTradingParameters } from "../../../constants/types";
 import { tradingParameters } from "../../../constants/trading-parameters.const";
 import { useSelector } from "react-redux";
 import { RootState } from "../../../store/types/roote-state.type";
-import { chooseСurrentVisibleTrading } from "../../../store/slices/trading.slice";
+import { chooseСurrentVisibleTrading } from "../../../features/tradings/slices/tradings.slice";
 import { useFetchService } from "../../../pages/traiding-page/hooks/useFetchService.hook";
 
 export const TableBody = () => {
