+++ src/app/features/tradings/slices/tradings.slice.ts
