+++ src/app/features/tradings/trading-navbar/types/tradings-data.type.ts
