+++ src/app/pages/traiding-page/traiding-page.tsx
@@ -1,16 +1,17 @@
 import "../traiding-page/traiding-page.scss";
-import { TradingNavbar } from "../../features/trading-navbar/trading-navbar";
+import { TradingNavbar } from "../../features/tradings/trading-navbar/trading-navbar";
 import { TableContainer } from "../../components/table/table-container";
 import { useFetchService } from "../../pages/traiding-page/hooks/useFetchService.hook";
 import { useDispatch, useSelector } from "react-redux";
-import { chooseСurrentVisibleTrading } from "../../store/slices/trading.slice";
+import { chooseСurrentVisibleTrading } from "../../features/tradings/slices/tradings.slice";
 import { tradingParametr } from "../../functions/index";
-import{changeParticipant} from"../../store/slices/table.slice";
+import{changeParticipant} from"../../features/participants/slices/participants.slice";
 
 export const TraidingPage = () => {
   const tradingData: any = useFetchService("http://localhost:3001/tradings");
   const activeTradingSelector = useSelector(chooseСurrentVisibleTrading);
 const dispatchChangeParticipantsList=useDispatch()
+
   return (
     <>
       <TradingNavbar />
