+++ src/app/store/index.ts
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
-import activeParticipantReducer from "./slices/table.slice";
-import  activeTradingReducer  from "./slices/trading.slice";
+import activeParticipantReducer from "../features/participants/slices/participants.slice";
+import  activeTradingReducer  from "../features/tradings/slices/tradings.slice";
 export const store= configureStore({
     reducer:{
         activeParticipant:activeParticipantReducer,
