--- src/app/store/slices/table.slice.ts
@@ -1,18 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-import { currentActiveParticipant } from "../../functions/current-active-participant.funс";
-
-
-const activeParticipantReducer = createSlice({
-  name: "activeParticipant",
-  initialState: {
-    activeParticipant: currentActiveParticipant(),
-  },
-  // В reducers перечисляется набор методов, который затем будем использовать
-  reducers: {
-    changeParticipant:(state,action) =>{
-      state.activeParticipant = currentActiveParticipant(action.payload);
-    },
-  },
-});
-export const { changeParticipant } = activeParticipantReducer.actions;
-export default activeParticipantReducer.reducer;
